sahur_poc/sahur/sahur-mcp/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
websockets
httpx
python-dotenv>=1.0.0
# sahur-core는 개발 환경에서 pip install -e ../sahur-core로 설치
